started work on turrets
started work on plane turrets. Ideally these will be handled entirely by code and feature a moving sprite that always stays relative to the plane body, and it can rotate and shoot on its own. Also fixed a minor issue with the player's hovering mode which caused their velocity to rapidly oscillate around 0.